Use Markdown-style quoting
Interesting side note: this is a sermon I wrote almost 15 years ago, so the formatting standard I used back then isn't _exactly_ what we use nowadays any more. ;) Seriously, the Markdown format _didn't even exist_ when this text was written.

So, doing a bit of refreshing of it to look better when formatted as Markdown today.