Move includes folder to _includes
This is a preparational step for getting the blog post support added,
and also for automating the whole build process. The whole `Makefile`
setup we are using now is quite primitive, to be honest. It would be
much better to automate this, and in the process of doing so, it makes
sense to be able to exclude certain folders from automatic compilation.

Hence, I am now introducing something I'm not overly fond of:
_convention_ over configuration. :) Yes, we could just add a blacklist
of excluded folders in `config.toml` but conventions might not be
*entirely* bad either. ;) Why not have both? Conventions for the naming,
and still being able to explicitly specifying it in the `config.toml`,
for maximum control.